Use long double as return type when accessing D(,) from python

+++ plots.py
@@ -23,7 +23,7 @@ c_D_roots = c_functions.find_k_perp_roots_array
 # c_D_array.restype = None
 
 c_D.argtypes = (ctypes.c_double, ctypes.c_double)
-c_D.restype = ctypes.c_double
+c_D.restype = ctypes.c_longdouble
 
 c_D_roots.argtypes = (ctypes.POINTER(ctypes.c_double), ctypes.POINTER(ctypes.c_double), ctypes.POINTER(ctypes.c_double), ctypes.c_int)
 c_D_roots.restype = ctypes.c_int
